Move normalizers to a separate file
